获取合伙人的成交单

+++ src/main/java/com/dazhi/authdemo/modules/auth/entity/DealEntity.java
@@ -23,7 +23,7 @@ public class DealEntity {
     // 经销商id
     private Long jsxAccountId;
     // 三工合伙人id
-    private Integer sgAccountId;
+    private Long sgAccountId;
 
     // 产品
     private String productName;
